requirements.txt
Django==5.2.9
djangorestframework==3.16.1
openpyxl==3.1.5
